blog/search_indexes.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-
"""
@Time    : 2020/6/2 10:29
@File    : search_indexes.py
@author  : dfkai
@Software: PyCharm
"""

from haystack import indexes

from .models import Article


class ArticleIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    def get_model(self):
        return Article

    def index_queryset(self, using=None):
        return self.get_model().objects.filter(status='p')
