requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.6
django-elasticsearch-dsl==7.1.1
django-haystack==2.8.1
django-mdeditor==0.1.18
django-suit==2.0a1
django-uuslug==1.2.0
elasticsearch==7.7.1
elasticsearch-dsl==7.2.1
jieba==0.42.1
mistune==0.8.4
Pygments==2.6.1
python-dateutil==2.8.1
python-slugify==4.0.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.25.9
Whoosh==2.7.4
